1.修正x64编译失败的BUG 2.增加内存池的性能检测Case 3.内存池性能分析报告
git-svn-id: http://qlanguage.googlecode.com/svn/trunk@297 6d982bb3-a5a7-393d-ffc5-582415be9c85

+++ Source/QCore/Library/callstack.cpp
@@ -93,7 +93,7 @@ void CallStack::getFuncInfo(LPVOID dwFunc, FuncInfo& info)
 
     DWORD64 dwDisplacement = 0;
 
-    if (SymGetSymFromAddr64(hProcess, dwFunc, &dwDisplacement, symbol))
+    if (SymGetSymFromAddr64(hProcess, (DWORD)dwFunc, &dwDisplacement, symbol))
     {
         strncpy(info.szFuncName, symbol->Name, min(sizeof(info.szFuncName) - 1, strlen(symbol->Name)));
         info.szFuncName[min(sizeof(info.szFuncName) - 1, strlen(symbol->Name))] = 0;
@@ -102,7 +102,7 @@ void CallStack::getFuncInfo(LPVOID dwFunc, FuncInfo& info)
     IMAGEHLP_LINE64 imageHelpLine;
     imageHelpLine.SizeOfStruct = sizeof(imageHelpLine);
 
-    if (SymGetLineFromAddr64(hProcess, dwFunc, (PDWORD)&dwDisplacement, &imageHelpLine))
+    if (SymGetLineFromAddr64(hProcess, (DWORD)dwFunc, (PDWORD)&dwDisplacement, &imageHelpLine))
     {
         strncpy(info.szFilePath, imageHelpLine.FileName, min(sizeof(info.szFilePath) - 1, strlen(imageHelpLine.FileName)));
         info.szFilePath[min(sizeof(info.szFilePath) - 1, strlen(imageHelpLine.FileName))] = 0;
